Fix Maven warnings
Maven issued several warnings when running the project. In most cases,
these were just deprecation warnings like switching from ${name} to
${project.name} or including the report plugin with its version number.

In one case, a plugin configuration related to Eclipse couldn't be
resolved. Since this pom.xml is targeted to be pure in the way that
there are no IDE specific settings included, this part was excluded.