Set Java source and target version to 1.8
This change was made to prepare compilation with the Java compiler of
JDK9, which doesn't support versions 1.5 or below anymore. Since Java
version 1.8 is the only other freely available edition still in support
by Oracle, this setting should suffice.

One additional bit of information: To actually cross-compile correctly,
javac needs the option -bootclasspath set, so it is able to find the
Java libraries. Actual cross-compiling in Maven should be configured
using the toolchain-plugin.