Remove distribution management and repositories
Since no artifacts are published on a public repository like Maven
Central, forcing a certain distribution management configuration doesn't
make sense. This may change in the future due to a definition of a
publicly accessible location where artifacts maybe pushed to.

The same thoughts led to removal of the repositories configuration.
Maven uses the user defaults for now.